finalized stuff, ready for port/final testing

+++ src/main/java/twopiradians/minewatch/common/entity/EntityMcCreeBullet.java
@@ -44,7 +44,7 @@ public class EntityMcCreeBullet extends EntityMWThrowable {
 		
 		if (result.entityHit instanceof EntityLivingBase && this.getThrower() instanceof EntityPlayer &&
 				result.entityHit != this.getThrower() && !this.world.isRemote) {
-			float damage = 70 - (70 - 21) * (this.ticksExisted / lifetime);
+			float damage = 70 - (70 - 21) * ((float)this.ticksExisted / lifetime);
 			((EntityLivingBase)result.entityHit).attackEntityFrom(DamageSource.causePlayerDamage((EntityPlayer) this.getThrower()), damage/ItemMWWeapon.DAMAGE_SCALE);
 			((EntityLivingBase)result.entityHit).hurtResistantTime = 0;
 			result.entityHit.world.playSound(null, this.getThrower().posX, this.getThrower().posY, this.getThrower().posZ,
