+++ src/main/java/twopiradians/minewatch/common/entity/EntitySoldier76Bullet.java
@@ -44,7 +44,7 @@ public class EntitySoldier76Bullet extends EntityMWThrowable {
 
 		if (result.entityHit instanceof EntityLivingBase && this.getThrower() != null &&
 				result.entityHit != this.getThrower() && !this.world.isRemote) {
-			float damage = 19 - (19 - 5.7f) * (this.ticksExisted / lifetime);
+			float damage = 19 - (19 - 5.7f) * ((float)this.ticksExisted / lifetime);
 			((EntityLivingBase)result.entityHit).attackEntityFrom(DamageSource.causePlayerDamage((EntityPlayer) this.getThrower()), damage/ItemMWWeapon.DAMAGE_SCALE);
 			((EntityLivingBase)result.entityHit).hurtResistantTime = 0;
 			result.entityHit.world.playSound(null, this.getThrower().posX, this.getThrower().posY, this.getThrower().posZ,
